+++ src/ResourceManager/AnalysisServices/Commands.AnalysisServices.Dataplane/Models/TokenCacheItemProvider.cs
@@ -39,8 +39,18 @@ namespace Microsoft.Azure.Commands.AnalysisServices.Dataplane.Models
 
         public string GetTokenFromTokenCache(TokenCache tokenCache, string uniqueId, string environmentName)
         {
-            var tokenCacheItems = tokenCache.ReadItems().Where(tokenItem => tokenItem.UniqueId.Equals(uniqueId) && (string.IsNullOrEmpty(environmentName) || tokenItem.Resource.Contains(environmentName)));
-            var tokenCacheItem = tokenCacheItems.FirstOrDefault();
+            TokenCacheItem tokenCacheItem = null;
+            if (string.IsNullOrEmpty(uniqueId))
+            {
+                var tokenCacheItems = tokenCache.ReadItems().Where(tokenItem => (string.IsNullOrEmpty(environmentName) || tokenItem.Resource.Contains(environmentName)));
+                tokenCacheItem = tokenCacheItems.FirstOrDefault();
+            }
+            else
+            {
+                var tokenCacheItems = tokenCache.ReadItems().Where(tokenItem => tokenItem.UniqueId.Equals(uniqueId) && (string.IsNullOrEmpty(environmentName) || tokenItem.Resource.Contains(environmentName)));
+                tokenCacheItem = tokenCacheItems.FirstOrDefault();
+            }
+
             if (tokenCacheItem == null || string.IsNullOrEmpty(tokenCacheItem.AccessToken))
             {
                 throw new PSInvalidOperationException(string.Format(Resources.NotLoggedInMessage, ""));