+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery/ReplicationProtectedItem/GetAzureRmRecoveryServicesAsrReplicationProtectedItem.cs
@@ -34,7 +34,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
     public class GetAzureRmRecoveryServicesAsrReplicationProtectedItem : SiteRecoveryCmdletBase
     {
         /// <summary>
-        ///     Gets or sets Name of the Replicated Protected Item.
+        ///     Gets or sets the name of the replication protected item to get.
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByObjectWithName,
@@ -43,7 +43,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         public string Name { get; set; }
 
         /// <summary>
-        ///     Gets or sets friendly name of the Protection Entity.
+        ///     Gets or sets the friendly name of the replication protected item to get.
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByObjectWithFriendlyName,
@@ -52,7 +52,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         public string FriendlyName { get; set; }
 
         /// <summary>
-        ///     Gets or sets Protection Container Object.
+        ///     Gets or sets the ASR protection container object of the ASR protection container corresponding
+        ///     to the replication protected item. 
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByObject,
@@ -70,7 +71,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         public ASRProtectionContainer ProtectionContainer { get; set; }
 
         /// <summary>
-        ///     Gets or sets Protectable Item.
+        ///     Gets or sets an ASR protectable item object. The cmdlet gets the ASR replication protected item corresponding 
+        ///     to the specified ASR protectable item if the item is protected.
         /// </summary>
         [Parameter(
             ParameterSetName = ASRParameterSets.ByProtectableItemObject,